refactor: centralize template types
